==> rpdr_data_wrangling/4_asthma_exacerbations.Rmd <==
---
title: "PEGASUS Asthma Exacerbations"
author: "MHuang"
date: "8/20/2020"
output: 
  html_document: 
    toc: yes
    toc_float: 
      collapsed: false
      smooth_scroll: false
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}

## Packages

pkg <- c("tidyverse", "data.table", "here", "DT", "grid", "gridExtra")
for (p in pkg) {
        if (require(p, character.only = T)) {
                print(paste0(p, " loaded successfully"))
        } else {
                install.packages(p)
                require(p, character.only = T)
                print(paste0(p, " downloaded and loaded successfully"))
        }
}


## Paths

pegasus_dir <- "C:/Users/MHuang/Dropbox (Partners HealthCare)/PEGASUS/"
dat_dir <- "data/"
long_fmt_dir <- "data/long_format/"


## Filenames

rpdr_file_header <- str_c(pegasus_dir, "rpdr_data/rpdr_pull_082020/mns43_083120164337973193_")
rpdr_file_ending <- ".txt" 
all_subj_fname <- here(dat_dir, "all_subj_enc_bmi.csv")

```

# 1. Diagnoses data
## 1.1. Read in data & parse out asthma exacerbations

```{r}

all_subj <- fread(all_subj_fname)
all_subj[, Collect_Date := as.Date(Collect_Date, format = "%Y-%m-%d")]
all_subj[, .N, .(is.na(EMPI), is.na(Collect_Date))][order(is.na, is.na.1)] %>% 
        setnames(., c("missing_EMPI", "missing_Collect_Date", "N")) %>% print()

dia <- fread(str_c(rpdr_file_header, "Dia", rpdr_file_ending))
dia[, unique(EMPI)] %in% all_subj[, EMPI] %>% table()
dia[, Date := as.Date(Date, format = "%m/%d/%Y")]

dia <- merge(dia, all_subj[, .(EMPI, Collect_Date)], by = "EMPI", all.x = T)
dia[, dia_timegap_d := as.numeric(Collect_Date - Date)]
dia[, dia_timegap_y := dia_timegap_d/365.25]
dia[, summary(dia_timegap_y)]

dia_condensed <- dia %>% group_by(Diagnosis_Name, Code_Type, Code) %>% summarise() %>% as.data.table()
dia_condensed[grep("asthma", Diagnosis_Name, ignore.case = T), ] %>% datatable()
dia_condensed[grep("asthma", Diagnosis_Name, ignore.case = T), ][grep("exacer|asthmaticus", Diagnosis_Name, ignore.case = T), ] %>% 
        datatable()

```

## 1.2. Filter out events after serum collection

This will effectively exclude subjects with no serum collection date

Status asthmaticus is a medical emergency, an extreme form of asthma exacerbation characterized by hypoxemia, hypercarbia, and secondary respiratory failure (Ref: [NCBI Book](https://www.ncbi.nlm.nih.gov/books/NBK526070/)). 

```{r}

dia_trim <- dia[Date <= Collect_Date, ]
dia_trim[, length(unique(EMPI))] 
dia_trim[, unique(EMPI)] %in% all_subj[SampleGroup == "Asthmatic", EMPI] %>% table()

asth_exac_trim <- dia_trim[grep("asthma", Diagnosis_Name, ignore.case = T), 
                           ][grep("exacer|with status asthmaticus", Diagnosis_Name, ignore.case = T), ]


## There may be multiple records for the same exacerbation on the same day

asth_exac_trim_uniq <- unique(asth_exac_trim[, .(EMPI, Date, Diagnosis_Name, Code_Type, Code, 
                                                 Collect_Date, dia_timegap_d, dia_timegap_y)])
asth_exac_trim_uniq[, asthma_exacerbation := 1]
asth_exac_trim_uniq[grep("with status asthmaticus", Diagnosis_Name, ignore.case = T), status_asthmaticus := 1]
asth_exac_trim_uniq[, summary(dia_timegap_y)]

```

## 1.3. Create variables for number of exacerbations (total and total over a period of years)

```{r}

asth_exac_smry <- asth_exac_trim_uniq[, .(asthma_exacerbation_trim_totnum = sum(asthma_exacerbation, na.rm = T), 
                                          status_asthmaticus_trim_totnum = sum(status_asthmaticus)), by = EMPI]
asth_exac_smry_1y <- asth_exac_trim_uniq[dia_timegap_y <= 1, 
                                         .(asthma_exacerbation_trim_totnum_1y = sum(asthma_exacerbation, na.rm = T), 
                                           status_asthmaticus_trim_totnum_1y = sum(status_asthmaticus)), by = EMPI]
asth_exac_smry <- merge(asth_exac_smry, asth_exac_smry_1y, by = "EMPI", all = T)

asth_exac_smry_2y <- asth_exac_trim_uniq[dia_timegap_y <= 2, 
                                         .(asthma_exacerbation_trim_totnum_2y = sum(asthma_exacerbation, na.rm = T), 
                                           status_asthmaticus_trim_totnum_2y = sum(status_asthmaticus)), by = EMPI]
asth_exac_smry <- merge(asth_exac_smry, asth_exac_smry_2y, by = "EMPI", all = T)

asth_exac_smry_3y <- asth_exac_trim_uniq[dia_timegap_y <= 3, 
                                         .(asthma_exacerbation_trim_totnum_3y = sum(asthma_exacerbation, na.rm = T), 
                                           status_asthmaticus_trim_totnum_3y = sum(status_asthmaticus)), by = EMPI]
asth_exac_smry <- merge(asth_exac_smry, asth_exac_smry_3y, by = "EMPI", all = T)

asth_exac_smry_5y <- asth_exac_trim_uniq[dia_timegap_y <= 5, 
                                         .(asthma_exacerbation_trim_totnum_5y = sum(asthma_exacerbation, na.rm = T), 
                                           status_asthmaticus_trim_totnum_5y = sum(status_asthmaticus)), by = EMPI]
asth_exac_smry <- merge(asth_exac_smry, asth_exac_smry_5y, by = "EMPI", all = T)

asth_exac_smry_10y <- asth_exac_trim_uniq[dia_timegap_y <= 10, 
                                          .(asthma_exacerbation_trim_totnum_10y = sum(asthma_exacerbation, na.rm = T), 
                                            status_asthmaticus_trim_totnum_10y = sum(status_asthmaticus)), by = EMPI]
asth_exac_smry <- merge(asth_exac_smry, asth_exac_smry_10y, by = "EMPI", all = T)


## Set NA to 0

exac_cnt_cols <- grep("_trim_totnum", colnames(asth_exac_smry), value = T)
exac_cnt_cols
asth_exac_smry[, (exac_cnt_cols) := lapply(.SD, function(x) {replace(x, is.na(x), 0)}), .SDcols = exac_cnt_cols]
asth_exac_smry <- merge(asth_exac_smry, all_subj[, .(EMPI, SampleGroup)], by = "EMPI")

```

## 1.4. Add an indicator for exacerbation on day of serum collection

```{r}

asth_exac_trim_uniq[dia_timegap_d == 0, length(unique(EMPI))]
asth_exac_smry[EMPI %in% asth_exac_trim_uniq[dia_timegap_d == 0, unique(EMPI)], .N, SampleGroup]
asth_exac_smry[, exacerbation_on_serum_date := case_when(EMPI %in% asth_exac_trim_uniq[dia_timegap_d == 0, unique(EMPI)] ~ 1, 
                                                         TRUE ~ 0)]

```

# 2. Plot asthma exacerbation counts

```{r}

asth_exac_smry[SampleGroup == "Asthmatic", summary(asthma_exacerbation_trim_totnum)]
asth_exac_smry[SampleGroup == "Asthmatic", summary(status_asthmaticus_trim_totnum)]

asth_exac_smry[SampleGroup == "Control", summary(asthma_exacerbation_trim_totnum)]
asth_exac_smry[SampleGroup == "Control", summary(status_asthmaticus_trim_totnum)]
# dia_trim[EMPI %in% asth_exac_smry[SampleGroup == "Control" & asthma_exacerbation_trim_totnum > 0, EMPI], ]
# This subject only has one extrinsic asthma with exacerbation diagnosis >10yrs before (allergic rhinitis due to pollen diagnosis at the same time); will exclude 
asth_exac_smry <- asth_exac_smry[SampleGroup == "Asthmatic", ]


## asthma_exacerbation

g1 <- ggplot(asth_exac_smry[SampleGroup == "Asthmatic", ]) + 
        geom_histogram(aes(asthma_exacerbation_trim_totnum_1y), bins = 100) + 
        labs(x = "1yr before") + 
        ylim(c(0, 170)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g2 <- ggplot(asth_exac_smry[SampleGroup == "Asthmatic", ]) + 
        geom_histogram(aes(asthma_exacerbation_trim_totnum_2y), bins = 100) + 
        labs(x = "2yrs before") + 
        ylim(c(0, 170)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g3 <- ggplot(asth_exac_smry[SampleGroup == "Asthmatic", ]) + 
        geom_histogram(aes(asthma_exacerbation_trim_totnum_3y), bins = 100) + 
        labs(x = "3yrs before") + 
        ylim(c(0, 170)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g4 <- ggplot(asth_exac_smry[SampleGroup == "Asthmatic", ]) + 
        geom_histogram(aes(asthma_exacerbation_trim_totnum_5y), bins = 100) + 
        labs(x = "5yrs before") + 
        ylim(c(0, 170)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g5 <- ggplot(asth_exac_smry[SampleGroup == "Asthmatic", ]) + 
        geom_histogram(aes(asthma_exacerbation_trim_totnum_5y), bins = 100) + 
        labs(x = "10yrs before") + 
        ylim(c(0, 170)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
g6 <- ggplot(asth_exac_smry[SampleGroup == "Asthmatic", ]) + 
        geom_histogram(aes(asthma_exacerbation_trim_totnum), bins = 100) + 
        labs(x = "Entire time in system") + 
        ylim(c(0, 170)) + 
        theme_minimal() + 
        theme(axis.title = element_text(size = 12), 
              axis.text = element_text(size = 12))
grid.arrange(g1, g2, g3, g4, g5, g6, nrow = 2, 
             top = textGrob("Number of asthma acute exacerbations before/on serum collection", 
                            gp = gpar(fontsize = 16)))

```

# 3. Detailed tabulated data

```{r, eval = F}

datatable(asth_exac_smry)

asth_exac_smry[, .N, asthma_exacerbation_trim_totnum_1y][order(asthma_exacerbation_trim_totnum_1y)] %>% datatable()
asth_exac_smry[, .N, asthma_exacerbation_trim_totnum_2y][order(asthma_exacerbation_trim_totnum_2y)] %>% datatable()
asth_exac_smry[, .N, asthma_exacerbation_trim_totnum_3y][order(asthma_exacerbation_trim_totnum_3y)] %>% datatable()
asth_exac_smry[, .N, asthma_exacerbation_trim_totnum_5y][order(asthma_exacerbation_trim_totnum_5y)] %>% datatable()
asth_exac_smry[, .N, asthma_exacerbation_trim_totnum_10y][order(asthma_exacerbation_trim_totnum_10y)] %>% datatable()
asth_exac_smry[, .N, asthma_exacerbation_trim_totnum][order(asthma_exacerbation_trim_totnum)] %>% datatable()

```

# 4. Save data

```{r}

# Save data table with aggregated info & one row per subject

fwrite(asth_exac_smry, file = here(dat_dir, "asthma_exacerbation_summary.csv"))


## Save long format data

fwrite(asth_exac_trim_uniq, file = here(long_fmt_dir, "measthma_exacerbation_trim_long.csv"))

```

# 5. Session info

```{r}

sessionInfo()

```
